+++ @types/client.d.ts
@@ -1,14 +1,17 @@
-import { Account, Order, Position, Asset, Watchlist, Calendar, Clock, AccountConfigurations, PortfolioHistory, Activity, DefaultCredentials, OAuthCredentials, OrderCancelation, PageOfTrades, PageOfQuotes, PageOfBars, Bar_v1, LastQuote_v1, LastTrade_v1, Snapshot, NewsPage, LatestTrade } from './entities.js';
+import { Account, Order, Position, Asset, Watchlist, Calendar, Clock, AccountConfigurations, PortfolioHistory, Activity, DefaultCredentials, OAuthCredentials, OrderCancelation, PageOfTrades, PageOfQuotes, PageOfBars, Bar_v1, LastQuote_v1, LastTrade_v1, Snapshot, NewsPage, LatestTrade, Endpoints } from './entities.js';
 import { GetOrder, GetOrders, PlaceOrder, ReplaceOrder, CancelOrder, GetPosition, ClosePosition, GetAsset, GetAssets, GetWatchList, CreateWatchList, UpdateWatchList, AddToWatchList, RemoveFromWatchList, DeleteWatchList, GetCalendar, UpdateAccountConfigurations, GetAccountActivities, GetPortfolioHistory, GetBars, GetBars_v1, GetTrades, GetQuotes, GetLastTrade_v1, GetLastQuote_v1, GetSnapshot, GetSnapshots, ClosePositions, GetNews, GetLatestTrade } from './params.js';
 export declare class AlpacaClient {
     params: {
-        credentials?: DefaultCredentials | OAuthCredentials;
         rate_limit?: boolean;
+        endpoints?: Endpoints | Map<keyof Endpoints, any>;
+        credentials?: DefaultCredentials | OAuthCredentials;
     };
+    private baseURLs;
     private limiter;
     constructor(params: {
-        credentials?: DefaultCredentials | OAuthCredentials;
         rate_limit?: boolean;
+        endpoints?: Endpoints | Map<keyof Endpoints, any>;
+        credentials?: DefaultCredentials | OAuthCredentials;
     });
     isAuthenticated(): Promise<boolean>;
     getAccount(): Promise<Account>;