+++ @types/entities.d.ts
@@ -1,3 +1,15 @@
+export interface Endpoints {
+    rest: {
+        beta: 'https://data.alpaca.markets/v1beta1';
+        account: 'https://api.alpaca.markets/v2';
+        market_data_v2: 'https://data.alpaca.markets/v2';
+        market_data_v1: 'https://data.alpaca.markets/v1';
+    };
+    websocket: {
+        account: 'wss://api.alpaca.markets/stream';
+        market_data: (source: DataSource) => string;
+    };
+}
 /**
  * Your Alpaca key id and secret.
  * Can be passed to the AlpacaClient and AlpacaStream.