+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [6.3.19](https://github.com/117/alpaca/compare/v6.3.18...v6.3.19) (2022-06-10)
 
+- Add support for custom endpoint urls (#105)
+
 ### [6.3.18](https://github.com/117/alpaca/compare/v6.3.17...v6.3.18) (2022-05-09)
 
 ### [6.3.17](https://github.com/117/alpaca/compare/v6.3.16...v6.3.17) (2022-04-06)